# Parallelize precompiles generation