This PR parallelizes `generate_precompile.jl`. It saves half a minute building Julia, and, thanks to using `Channel`, the order of the precompiles should be still fully deterministic.

**PR**
```
...
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 40/40
Executing precompile statements... 1750/1754
Precompilation complete. Summary:
Total ─────── 112.036051 seconds
...
```


**Master**
```
...
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 40/40
Executing precompile statements... 1748/1752
Precompilation complete. Summary:
Generation ── 121.136666 seconds 80.6719%
Execution ───  29.023100 seconds 19.3281%
Total ─────── 150.159765 seconds
...
```

The next optimization would be to read `precompile_file` on the fly, that would save several seconds extra.