@IanButterworth 2022-12-26
Seems like a good thing to do, but it could make the precompilation process harder to debug. It might be good to be able to fallback to the current serial approach for debugging purposes. I guess putting optional waits on the `@async` tasks could be a simple way to do that?

Also I think the progress printing needs figuring out. It's currently a bit chaotic and hard to interpret.