@petvana 2022-12-27
I've tried to improve the printing (link to asciinema below) and add `PARALLEL_PRECOMPILATION = false` option for serial debugging 

<a href="https://asciinema.org/a/548119?autoplay=1"><img src="https://asciinema.org/a/548119.png" width="600"/></a>