@petvana 2022-12-28
> The REPL precomp generation script includes intentionally throwing errors. Perhaps that's why errormonitor etc was surfacing errors that were previously ignored?

Perhaps, but I'm unable to confirm locally.