@petvana 2022-12-29
Besides `errormonitor`, the PR seems ready. Suggestions for improving the progress printing welcomed.