@IanButterworth 2022-12-29
Looks great. 
For formatting maybe this, where `Waiting` etc. are colored and the counts for the categories are colored appropriately. Also perhaps "Basic" instead of "normal"?
```
Precompile statements (Waiting, Running, Finished): 
└ Collect 1748 (Basic: 729, REPL 40/40: 1653) => Execute 1744/1748
Precompilation complete. Summary:
Total ─────── 114.175765 seconds
Outputting sysimage file...
Output ────── 162.738189 seconds
````