@petvana 2022-12-29
New fancy printing. Running for `sys.so` to make it short.

<a href="https://asciinema.org/a/548532" target="_blank"><img src="https://asciinema.org/a/548532.svg" width=400 /></a>