@KristofferC 2022-12-29
I like it!

The raw logs in e.g. CI gets a bit spammed 

```
_bk;t=1672318930123Collecting and executing precompile statements
_bk;t=1672318930123
└ Collect (Basic: ? , REPL 0/0: ? ) => Execute ? 
└ Collect (Basic: ? , REPL 0/0: ? ) => Execute ? 
└ Collect (Basic: ◐ , REPL 0/0: ? ) => Execute ? 
└ Collect (Basic: ◐ , REPL 0/0: ◐ ) => Execute ? 
└ Collect (Basic: ◐ , REPL 0/0: ◐ ) => Execute ◐ 1 (1 failed)
└ Collect (Basic: ◓ , REPL 0/0: ◓ ) => Execute ◓ 1 (1 failed)
└ Collect (Basic: ◓ , REPL 0/0: ◓ ) => Execute ◓ 2 (1 failed)
└ Collect (Basic: ◓ , REPL 0/0: ◓ ) => Execute ◓ 3 (3 failed)
└ Collect (Basic: ◓ , REPL 0/0: ◓ ) => Execute ◓ 4 (3 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 4 (3 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 5 (3 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 6 (3 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 7 (4 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 8 (4 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 9 (4 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 10 (4 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 11 (4 failed)
└ Collect (Basic: ◑ , REPL 0/0: ◑ ) => Execute ◑ 12 (4 failed)
└ Collect (Basic: ◒ , REPL 0/0: ◒ ) => Execute ◒ 12 (4 failed)
└ Collect (Basic: ◒ , REPL 0/0: ◒ ) => Execute ◒ 13 (4 failed)
└ Collect (Basic: ◒ , REPL 0/0: ◒ ) => Execute ◒ 14 (4 failed)
└ Collect (Basic: ◒ , REPL 0/0: ◒ ) => Execute ◒ 15 (4 failed)
```

Maybe check if stdout/stderr is a TTY for fancyprint?