@KristofferC 2022-12-29
If we want to be consistent with `Pkg.precompile`, the spinner should just be white I think.