@IanButterworth 2022-12-29
Looking great. Disabling the cursor is best with a try-finally because it can be annoying to lose it if the build process fails