@IanButterworth 2022-12-29
I realized the 4 we're reporting as failures are just comments being skipped, so (I hope you don't mind) I pushed a fix for that