@IanButterworth 2022-12-29
> The next optimization would be to read precompile_file on the fly, that would save several seconds extra.

It seems like `--trace-compile=$precompile_file` is populated at the end of the process. Making that more parallel might take a lot of internals work