@KristofferC 2022-12-29
We could flush the strean for every statement printed maybe.