@IanButterworth 2022-12-29
I actually think I might be wrong, and we're already flushing. Investigating..