@petvana 2022-12-29
> It seems like `--trace-compile=$precompile_file` is populated at the end of the process. Making that more parallel might take a lot of internals work

Actually, they are populated during the process. I've already tried a very naive approach and it seems to work.
https://github.com/petvana/julia/compare/pv/parallel-repl-statements-v2...petvana:julia:pv/parallel-repl-statements-on-the-fly-v2