@IanButterworth 2022-12-29
Nice. I've been trying an approach that doesn't re-read and unique the statements each loop, though it's not working yet