@petvana 2022-12-29
I propose to leave the asynchronous reading to a different PR because it adds one more level of complexity (and space for possible dead-locks).