@IanButterworth 2022-12-29
This seems to work without re-reading the total file each time https://github.com/IanButterworth/julia/commit/5ee671ad3c40533a3b3965f8bcbce20d18014cbf?w=1