@petvana 2022-12-29
> This seems to work without re-reading the total file each time [IanButterworth@5ee671a?w=1](https://github.com/IanButterworth/julia/commit/5ee671ad3c40533a3b3965f8bcbce20d18014cbf?w=1)

Nice, but I'm slightly worried about `readline` if it may split the last line before it's fully written. I'd rather use `readavailable` and split manually.