@vtjnash 2022-12-29
readline will never do that (if reading from a proper stream/pipe), but readavailable will often do that