@KristofferC 2022-12-29
Wow, I didn't mean to merge this. I tried to merge it locally using `gh merge` and it merged it here instead 😆