@KristofferC 2022-12-29
I think it asked me and since I thought it was local I thought whatever. The last step said "submit" which in hind sight probably should have made me more alert...