@petvana 2022-12-29
> readline will never do that (if reading from a proper stream/pipe), but readavailable will often do that

It seems possible if someone flushes (that is a correct think to do). The following MWE
``` julia
t = tempname()
w = open(t, "w")
r = open(t, "r")
tr = @async begin
    for i = 1:30
        sleep(0.1)
        line = readline(r)
        line == "" && continue
        println(line)
    end
end

print(w, "a+")
flush(w)
sleep(0.3)
print(w, "b\n")
close(w)

wait(tr)
```
prints only 
```
a+
```
but it is not a full line.