@vtjnash 2022-12-30
A file is not a stream/pipe (unless it was created with mkfifo)